Separate out scoring functions to separate file
